Move log format functions to new module
